Added scraper-script option to control English rating/certification systems.
The old UKRating setting still works as originally intended to enable the UK/BFFC rating system, but has been repurposed to now focus on the British-English movie title option. Perhaps later the var-name can be changed, but for the time being in this transition period it will work fine as-is.